Move SplitContainerMessageFilter and SplitContainerTypedControlCollection to separate files
