+++ src/System.Windows.Forms/src/System/Windows/Forms/SplitContainer.cs
@@ -2501,86 +2501,5 @@ namespace System.Windows.Forms
         {
             return new SplitContainerTypedControlCollection(this, typeof(SplitterPanel), /*isReadOnly*/true);
         }
-
-        private class SplitContainerMessageFilter : IMessageFilter
-        {
-            private readonly SplitContainer _owner;
-
-            public SplitContainerMessageFilter(SplitContainer splitContainer)
-            {
-                _owner = splitContainer;
-            }
-
-            bool IMessageFilter.PreFilterMessage(ref Message m)
-            {
-                if (m.MsgInternal >= User32.WM.KEYFIRST && m.MsgInternal <= User32.WM.KEYLAST)
-                {
-                    if ((m.MsgInternal == User32.WM.KEYDOWN && (Keys)m.WParamInternal == Keys.Escape)
-                        || (m.MsgInternal == User32.WM.SYSKEYDOWN))
-                    {
-                        // Notify that splitMOVE was reverted. This is used in ONKEYUP.
-                        _owner._splitBegin = false;
-                        _owner.SplitEnd(false);
-                        _owner._splitterClick = false;
-                        _owner._splitterDrag = false;
-                    }
-
-                    return true;
-                }
-
-                return false;
-            }
-        }
-
-        /// <summary>
-        ///  This control collection only allows a specific type of control
-        ///  into the controls collection.  It optionally supports readonlyness.
-        /// </summary>
-        internal class SplitContainerTypedControlCollection : TypedControlCollection
-        {
-            private readonly SplitContainer _owner;
-
-            public SplitContainerTypedControlCollection(Control c, Type type, bool isReadOnly) : base(c, type, isReadOnly)
-            {
-                _owner = c as SplitContainer;
-            }
-
-            public override void Remove(Control value)
-            {
-                if (value is SplitterPanel)
-                {
-                    if (!_owner.DesignMode)
-                    {
-                        if (IsReadOnly)
-                        {
-                            throw new NotSupportedException(SR.ReadonlyControlsCollection);
-                        }
-                    }
-                }
-
-                base.Remove(value);
-            }
-
-            internal override void SetChildIndexInternal(Control child, int newIndex)
-            {
-                if (child is SplitterPanel)
-                {
-                    if (!_owner.DesignMode)
-                    {
-                        if (IsReadOnly)
-                        {
-                            throw new NotSupportedException(SR.ReadonlyControlsCollection);
-                        }
-                    }
-                    else
-                    {
-                        // just no-op it at DT.
-                        return;
-                    }
-                }
-
-                base.SetChildIndexInternal(child, newIndex);
-            }
-        }
     }
 }
